requirements.txt
# Web framework for headless server
fastapi>=0.104.0
uvicorn>=0.24.0

# Streamlit for development interface
streamlit>=1.28.0

# Core dependencies
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# HTTP clients and async support
httpx>=0.25.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Data processing
pyyaml>=6.0.0
python-dateutil>=2.8.0

# Database
alembic>=1.13.0

# Soil data
soilgrids>=0.1.4

# Web scraping for market prices
beautifulsoup4>=4.12.0
selenium>=4.15.0
flask>=2.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Logging
structlog>=23.2.0
